+++ src/app/admin/page.js
@@ -1,65 +1,83 @@
-"use client"
-import React, { useState } from 'react';
-import { MdOutlineSpaceDashboard } from "react-icons/md";
-import { SlCalender } from "react-icons/sl";
-import { FaKey } from "react-icons/fa";
-import { HiOutlineUsers } from "react-icons/hi2";
-import DashBoard from '@/components/admin/DashBoard';
-import Bookings from '@/components/admin/Bookings';
-import Hotels from '@/components/admin/Hotels';
-import Users from '@/components/admin/Users';
+"use client";
 
-const Page = () => {
-  const [currentView, setCurrentView] = useState('dashboard');
+import { useState } from 'react';
+import { useRouter } from 'next/navigation';
 
-  const renderComponent = () => {
-    switch (currentView) {
-      case 'dashboard':
-        return <DashBoard />;
-      case 'bookings':
-        return <Bookings />;
-      case 'hotels':
-        return <Hotels />;
-      case 'users':
-        return <Users />;
-      default:
-        return <Users />;
+const LoginPage = () => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [error, setError] = useState(null);
+  const router = useRouter();
+
+  const handleLogin = async (e) => {
+    e.preventDefault();
+
+    // Simulate login
+    if (email === 'admin@example.com' && password === 'password') {
+      router.push('/admin/dashboard');
+    } else {
+      setError('Invalid credentials');
     }
   };
 
   return (
-    <div className='flex font-f_3 min-h-screen bg-zinc-100 overflow-hidden'>
-      {/* aside nav bar */}
-      <div className='fixed h-full w-[25%] md:w-[15%] lg:w-[10%] flex flex-col shadow-lg items-center gap-10 bg-[#FFFFFF]'>
-        <h1 className='text-[20px] text-black my-4'>LHC</h1>
-        <MdOutlineSpaceDashboard
-          size={35}
-          className={`cursor-pointer ${currentView === 'dashboard' ? 'text-black' : 'text-gray-500'}`}
-          onClick={() => setCurrentView('dashboard')}
-        />
-        <SlCalender
-          size={30}
-          className={`cursor-pointer ${currentView === 'bookings' ? 'text-black' : 'text-gray-500'}`}
-          onClick={() => setCurrentView('bookings')}
-        />
-        <FaKey
-          size={30}
-          className={`cursor-pointer ${currentView === 'hotels' ? 'text-black' : 'text-gray-500'}`}
-          onClick={() => setCurrentView('hotels')}
-        />
-        <HiOutlineUsers
-          size={30}
-          className={`cursor-pointer ${currentView === 'users' ? 'text-black' : 'text-gray-500'}`}
-          onClick={() => setCurrentView('users')}
-        />
-      </div>
+    <div className="flex min-h-screen items-center justify-center bg-gray-100 ">
+      <div className="bg-white p-12 rounded-lg shadow-xl w-full max-w-lg">
+        {/* Logo or Branding */}
+        <div className="text-center mb-8 ">
+          <h1 className="text-xl  text-gray-800 mb-6">The Luxury Hotel <span className="text-yellow-500">Concierge</span></h1>
+          <p className="text-4xl font-bold text-black-600">Welcome back!</p>
+        </div>
 
-      {/* main content area */}
-      <div className='ml-[25%] md:ml-[15%] lg:ml-[10%] w-[75%] md:w-[85%] lg:w-[90%] h-full p-4'>
-        {renderComponent()}
+        <form onSubmit={handleLogin} className="space-y-6">
+          <div>
+            <label className="block text-sm font-medium text-gray-700">Email address</label>
+            <input 
+              type="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              required 
+              className="w-full mt-1 p-3 border border-gray-300 rounded-lg  "
+              placeholder="Enter your email"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">Password</label>
+            <input 
+              type="password" 
+              value={password} 
+              onChange={(e) => setPassword(e.target.value)} 
+              required 
+              className="w-full mt-1 p-3 border border-gray-300 rounded-lg "
+              placeholder="Enter your password"
+            />
+          </div>
+
+          {error && (
+            <div className="text-red-500 text-sm">
+              {error}
+            </div>
+          )}
+
+          <div className="flex justify-between items-center">
+            <div>
+              <input type="checkbox" id="remember-me" className="mr-2" />
+              <label htmlFor="remember-me" className="text-sm text-gray-600">Remember me</label>
+            </div>
+            <a href="#" className="text-sm text-yellow-500 hover:underline">Forgot password?</a>
+          </div>
+
+          <button 
+            type="submit" 
+            className="w-full py-3 px-6 bg-black text-white font-semibold rounded-lg hover:bg-gray-800 "
+          >
+            Log In
+          </button>
+        </form>
       </div>
     </div>
   );
-}
+};
 
-export default Page;
+export default LoginPage;
