+++ src/base_url.js
@@ -1,3 +1,3 @@
-export const base_url = "https://thlc-backend-beta.vercel.app";
+export const base_url = "https://thlc-backend.vercel.app";
 // export const base_url = "http://localhost:4000";
 // https://thlc-backend-beta.vercel.app/
